requirements.txt
requests>=2.31.0
dnspython>=2.4.2
python-whois>=0.8.0
beautifulsoup4>=4.12.2
pyyaml>=6.0.1
colorama>=0.4.6
tqdm>=4.66.1
shodan>=1.31.0


beautifulsoup4

